requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.4
amqp==5.1.1
annotated-types==0.6.0
appdirs==1.4.3
asgiref==3.7.2
astroid==2.3.3
async-timeout==4.0.3
attrs==19.3.0
auto-py-to-exe==2.41.0
Automat==0.8.0
backports.zoneinfo==0.2.1
billiard==4.1.0
blinker==1.4
bottle==0.12.25
bottle-websocket==0.2.9
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==23.1.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.6
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-controlcenter==0.3.2
django-etc==1.4.0
django-pkgconf==0.4.0
django-sitemessage==1.4.0
djangorestframework==3.14.0
Eel==0.16.0
entrypoints==0.3
filelock==3.0.12
frozenlist==1.4.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==16.10.1
iotop==0.6
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
magic-filter==1.0.12
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.4
netifaces==0.10.4
oauthlib==3.1.0
packaging==23.2
pefile==2023.2.7
pexpect==4.6.0
Pillow==10.1.0
prompt-toolkit==3.0.39
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
pydantic==2.3.0
pydantic-core==2.6.3
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==8.0.1
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==5.3.1
requests==2.31.0
requests-oauthlib==0.6.2
requests-unixsocket==0.2.0
requirements-detector==0.6
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.3
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
telebot==0.0.3
text-unidecode==1.3
Twisted==18.9.0
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.7
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.8
whichcraft==0.6.1
wrapt==1.11.2
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
